2.1.0: Reverted utf8. To complicated to fully fix for now. git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@1804 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ protocol/amcp/AMCPProtocolStrategy.cpp
@@ -156,7 +156,7 @@ AMCPCommandPtr AMCPProtocolStrategy::InterpretCommandString(const std::wstring& 
 	AMCPCommandPtr pCommand;
 	MessageParserState state = New;
 
-	CASPAR_LOG(info) << u8(message);
+	CASPAR_LOG(info) << message;
 
 	std::size_t tokensInMessage = TokenizeMessage(message, &tokens);
